feat: reusable tooltip
