+++ src/modules/project/pages/projectView/views/projectMainBody/components/GoalCompleted.tsx
@@ -3,8 +3,10 @@ import { Box, HStack, Text, VStack } from '@chakra-ui/react'
 import { useTranslation } from 'react-i18next'
 import { MdModeEdit } from 'react-icons/md'
 
+import { DollarIconCircled, SatoshiIconCircled } from '../../../../../../../components/icons'
 import { Body1, H3 } from '../../../../../../../components/typography'
 import { IconButtonComponent } from '../../../../../../../components/ui'
+import { Tooltip } from '../../../../../../../components/ui/Tooltip'
 import { ProjectGoal, ProjectGoalCurrency } from '../../../../../../../types'
 import { getFormattedDate } from '../../../../../../../utils'
 import { useCurrencyFormatter } from '../../../hooks/useCurrencyFormatter'
@@ -54,12 +56,28 @@ export const GoalCompleted = ({ goal, isEditing = false, onOpenGoalModal }: Prop
       >
         <VStack width="100%">
           <HStack display="flex" alignItems="flex-start" justifyContent="space-between" width="100%">
-            <Body1 bold>
-              {formattedAmountContributed}{' '}
-              <Text as="span" color="neutral.600" fontWeight={500}>
-                {goal.currency === ProjectGoalCurrency.Btcsat ? `(${usdAmount})` : `(${satsAmount})`}
-              </Text>
-            </Body1>
+            <HStack display="flex" justifyContent="flex-start">
+              <Body1 bold>
+                {formattedAmountContributed}{' '}
+                <Text as="span" color="neutral.600" fontWeight={500}>
+                  {goal.currency === ProjectGoalCurrency.Btcsat ? `(${usdAmount})` : `(${satsAmount})`}
+                </Text>
+              </Body1>
+              {goal.currency === ProjectGoalCurrency.Btcsat ? (
+                <Tooltip px={4} py={1} content={<Text fontSize={12}>{t('This goal is denominated in Bitcoin')}</Text>}>
+                  <SatoshiIconCircled />
+                </Tooltip>
+              ) : (
+                <Tooltip
+                  px={4}
+                  py={1}
+                  content={<Text fontSize={12}>{t('This goal is denominated in US Dollars')}</Text>}
+                >
+                  <DollarIconCircled />
+                </Tooltip>
+              )}
+            </HStack>
+
             <HStack display="flex" alignItems="center" justifyContent="flex-end" gap={2}>
               <CheckCircleIcon color={'primary.500'} />
               <Body1 bold>
